Fix race condition... "unsafe"ly :p
If two guys called queryGHCI at the same time, they could race for the
lazy output of GHCI, which made no one happy.

Fix it with an MVar. Again, since we have to have access to said MVar in
a function where I have no idea how to get it, we have to use the
unsafePerformIO hack.